chore: Move components into more appropriate directories
